# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OfferLoots GPT Rewards Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful

  # User
...
  /offers:
    get:
      operationId: getOffers
      tags: [offers]
      summary: List available offers
      parameters:
        - in: query
          name: network
          schema:
            type: string
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: device
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Offers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOffers"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get single offer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found
...
  /apikeys:
    get:
      operationId: getApiKeys
      tags: [apikeys]
      summary: List user API keys
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [apikeys]
      summary: Generate a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /apikeys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [apikeys]
      summary: Delete an API key
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Networks
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, email, role, status, balance, points, rank, referralCode, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, banned, pending]
        balance:
          type: number
        points:
          type: integer
        rank:
          type: string
        referralCode:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Offer:
      type: object
      required: [id, name, payout, network, status, category, device, countries, description, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        payout:
          type: number
        network:
          type: string
        networkId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, paused, deleted]
        category:
          type: string
        device:
          type: string
          enum: [all, mobile, desktop]
        countries:
          type: array
          items:
            type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        offerUrl:
          type: ["string", "null"]
        completions:
          type: integer
        createdAt:
          type: string
...
    PaginatedOffers:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ApiKey:
      type: object
      required: [id, name, key, isActive, usageCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        key:
          type: string
        isActive:
          type: boolean
        usageCount:
          type: integer
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    # Networks
...