# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OfferLoots GPT Rewards Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  balance: number;
  points: number;
  rank: string;
  referralCode: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}